Fix Node Properties by converting from f64
Made nodes accept f64 arguments and convert them to f32 internally so that node properties work.

+++ node-graph/gmath-nodes/src/lib.rs
@@ -464,26 +464,32 @@ fn dot_product(_: impl Ctx, vector_a: DVec2, vector_b: DVec2) -> f64 {
 
 /// Constructs a color from RGBA components. Clamped to 0 -> 1.
 #[node_macro::node(category("Color"))]
-fn construct_color_rgba(_: impl Ctx, _primary: (), red: f32, green: f32, blue: f32, alpha: f32) -> Color {
-	Color::from_rgbaf32_unchecked(red.clamp(0., 1.), green.clamp(0., 1.), blue.clamp(0., 1.), alpha.clamp(0., 1.))
+fn construct_color_rgba(_: impl Ctx, _primary: (), red: f64, green: f64, blue: f64, alpha: f64) -> Color {
+	Color::from_rgbaf32_unchecked(red.clamp(0., 1.) as f32, green.clamp(0., 1.) as f32, blue.clamp(0., 1.) as f32, alpha.clamp(0., 1.) as f32)
 }
 
 /// Constructs a color from HSLA components. Clamped to 0 -> 1.
 #[node_macro::node(category("Color"))]
-fn construct_color_hsla(_: impl Ctx, _primary: (), hue: f32, saturation: f32, lightness: f32, alpha: f32) -> Color {
-	Color::from_hsla(hue.clamp(0.0, 1.0), saturation.clamp(0., 1.), lightness.clamp(0., 1.), alpha.clamp(0., 1.))
+fn construct_color_hsla(_: impl Ctx, _primary: (), hue: f64, saturation: f64, lightness: f64, alpha: f64) -> Color {
+	Color::from_hsla(hue.clamp(0.0, 1.0) as f32, saturation.clamp(0., 1.) as f32, lightness.clamp(0., 1.) as f32, alpha.clamp(0., 1.) as f32)
 }
 
 /// Constructs a color from a single grayscale value. Clamped to 0 -> 1.
 #[node_macro::node(category("Color"))]
-fn construct_color_grayscale(_: impl Ctx, _primary: (), luminance: f32) -> Color {
-	Color::from_luminance(luminance.clamp(0., 1.))
+fn construct_color_grayscale(_: impl Ctx, _primary: (), luminance: f64) -> Color {
+	Color::from_luminance(luminance.clamp(0., 1.) as f32)
 }
 
 /// Constructs a color from CMYKA components. Clamped to 0 -> 1.
 #[node_macro::node(category("Color"))]
-fn construct_color_cmyka(_: impl Ctx, _primary: (), cyan: f32, magenta: f32, yellow: f32, key: f32, alpha: f32) -> Color {
-	Color::from_cmyka(cyan.clamp(0., 1.), magenta.clamp(0., 1.), yellow.clamp(0., 1.), key.clamp(0., 1.), alpha.clamp(0., 1.))
+fn construct_color_cmyka(_: impl Ctx, _primary: (), cyan: f64, magenta: f64, yellow: f64, key: f64, alpha: f64) -> Color {
+	Color::from_cmyka(
+		cyan.clamp(0., 1.) as f32,
+		magenta.clamp(0., 1.) as f32,
+		yellow.clamp(0., 1.) as f32,
+		key.clamp(0., 1.) as f32,
+		alpha.clamp(0., 1.) as f32,
+	)
 }
 
 #[cfg(test)]
